Changed to use from_utf8_lossy again (#4)
The reason I changed to this before was that I couldn't get it working with the `Cow` object. Today, I checked the problem again and found the solution: `into_owned`, which takes ownership of the borrowed data (by copying it). Wonderful!

Also fixed a linting issue I happened to see when reading the Travis output.