Added a mode which lists all tagged versions (#7)
* Added a mode which lists all tagged versions

This mode is triggered by running the program in this mode:

```
cargo run ~/git/chaos # Only one parameter: the repo name
```

This will read the list of tags from git, and use them to produce a reasonably-nice looking changelog based on the list of git commits.

* Bug fix: Remove initial v prefix in tags

This is overly simplistic and agressive, and probably won't cover all cases properly. But _for me_, it works rather well and makes the program more useful, so let's go with this for now.

* Use a better way to get the git tags

The problem with the previous appraoch was that it included a bit of "junk" in terms of origin/master etc., which meant that the most recent tag was often excluded => no changelog entry would be produced for it. This has now all been fixed, and the program works like a charm.

* Peer review: Be more consistent with where we print the newline.